set intro page to default and add canvas page
