+++ components/Inputs/LunchAppButton.tsx
@@ -11,22 +11,22 @@ import {
 import { MERGE_CANVAS_CONTRACT_ADDRESS } from '../../utils/constants';
 import MergeCanvasArtifact from '../../contracts/MergeCanvas.json';
 import { usePixelCanvasContext } from '../../contexts/PixelCanvasContext';
+import Link from 'next/link';
 
 export const LunchButton = () => {
-    const lunchApp = useCallback(async () => {}, []);
-
     return (
         <div className="flex-none">
-            <button
-                type="button"
-                className={clsx(
-                    'py-5 px-6 w-full bg-eth-gray/90 text-white font-bold rounded-xl shadow transition cursor-pointer',
-                    'hover:bg-eth-gray hover:shadow-lg'
-                )}
-                onClick={() => lunchApp()}
-            >
-                Lunch Dapp
-            </button>
+            <Link href="/canvas">
+                <button
+                    type="button"
+                    className={clsx(
+                        'py-5 px-6 w-full bg-eth-gray/90 text-white font-bold rounded-xl shadow transition cursor-pointer',
+                        'hover:bg-eth-gray hover:shadow-lg'
+                    )}
+                >
+                    Lunch Dapp
+                </button>
+            </Link>
         </div>
     );
 };
