+++ package.json
@@ -16,14 +16,15 @@
     "lodash": "^4.17.21",
     "next": "12.1.6",
     "panzoom": "^9.4.3",
+    "postcss": "^8.4.14",
     "react": "18.1.0",
     "react-colorful": "^5.6.1",
+    "react-device-detect": "^2.2.2",
     "react-dom": "18.1.0",
     "react-icons": "^4.4.0",
-    "wagmi": "^0.6.5",
-    "postcss": "^8.4.14",
     "tailwindcss": "^3.1.0",
-    "typescript": "^4.8.3"
+    "typescript": "^4.8.3",
+    "wagmi": "^0.6.5"
   },
   "devDependencies": {
     "@types/lodash": "^4.14.184",
